fix!: unsigned 128bit Amount (#1742)
Description
---
fix!: unsigned 128bit Amount
sets version 0.19.0

Motivation and Context
---
Having a signed 192-bit integer had downsides:
- required many checks peppered through the code to ensure that values
are non-negative,
 - poorer performance and,
 - more on-chain data.
 
`u128::MAX` is an astronomical value, and would hold the total ETH
supply (w/ 18 decimals) for hundreds of trillions of years.

This PR includes a precision amount that can be optionally compiled into
templates (feature flag) if 128-bit precision is insufficient. This uses
roughly/naively 3500 bytes in the template binary.

Ref #1496  

How Has This Been Tested?
---
Existing tests and manually

What process can a PR reviewer use to test or verify this change?
---


Breaking Changes
---

- [ ] None
- [x] Requires data directory to be deleted
- [ ] Other - Please specify

+++ crates/engine_types/src/resource_container.rs
@@ -136,7 +136,7 @@ impl ResourceContainer {
     pub fn unlocked_amount(&self) -> Amount {
         match self {
             Self::Fungible { amount, .. } => *amount,
-            Self::NonFungible { token_ids, .. } => Amount::new(token_ids.len()),
+            Self::NonFungible { token_ids, .. } => Amount::new(token_ids.len() as u128),
             Self::Confidential { revealed_amount, .. } => *revealed_amount,
             Self::Stealth { revealed_amount, .. } => *revealed_amount,
         }
@@ -156,7 +156,7 @@ impl ResourceContainer {
     pub fn locked_amount(&self) -> Amount {
         match self {
             Self::Fungible { locked_amount, .. } => *locked_amount,
-            Self::NonFungible { locked_token_ids, .. } => Amount::new(locked_token_ids.len()),
+            Self::NonFungible { locked_token_ids, .. } => Amount::new(locked_token_ids.len() as u128),
             Self::Confidential {
                 locked_revealed_amount, ..
             } => *locked_revealed_amount,
